web/requirements.txt
Flask==2.2.5
psycopg2-binary==2.9.7
pytest==7.4.0
coverage==7.2.4
codecov==2.1.0
gunicorn==20.1.0


